revert timer

+++ chaos/include/Sequence.hpp
@@ -88,27 +88,27 @@ namespace Chaos {
      * \brief Emulates holding down the gamapad input signal in the fully on position.
      * \param signal The input signal the console expects to receive
      * \param value The value of the input signal.
-     * \param hold_time Time to hold the signal in seconds.
+     * \param hold_time Time to hold the signal in microseconds.
      *
      * A value of 0 means that the default maximum will be used when emulating the signal press.
      * This is 1 for buttons and the d-pad, and the joystick maximum for other axes. For the hybrid
      * controls L2/R2, a non-zero value is passed to the axis portion of the signal. The button
      * portion is set to 1 regardless.
      */
-    void addHold(std::shared_ptr<ControllerInput> signal, short value, double hold_time);
+    void addHold(std::shared_ptr<ControllerInput> signal, short value, unsigned int hold_time);
     /**
      * \brief Turns off the gamapad input signal.
      * \param signal The input signal the console expects to receive
-     * \param hold_time Time to hold the signal in seconds.
+     * \param hold_time Time to hold the signal in microseconds.
      */
-    void addRelease(std::shared_ptr<ControllerInput> signal, double release_time);
+    void addRelease(std::shared_ptr<ControllerInput> signal, unsigned int release_time);
     /**
      * \brief Add a delay
      * \param delay Time to delay in seconds
      *
      */
-    //void addDelay(unsigned int delay);
-    void addDelay(double delay);
+    void addDelay(unsigned int delay);
+    //void addDelay(double delay);
 
     void addSequence(std::shared_ptr<Sequence> seq);
 
@@ -130,7 +130,7 @@ namespace Chaos {
     /**
      * \brief Return the nth step in the sequence of events
      * 
-     * \param step 
+     * \param sequenceTime 
      * \return true if sequence is done
      * \return false if sequence is still in progress
      */
